projeto1/pyproject.toml
[project]
name = "projeto1"
version = "0.1.0"
description = "Primeiro projeto em fastapi"
authors = [
    {name = "tiagobernard",email = "tiago@tiagobernardes.com.br"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi (>=0.115.12,<0.116.0)",
    "unicorn (>=2.1.3,<3.0.0)",
    "requests (>=2.32.3,<3.0.0)",
    "aiohttp (>=3.11.16,<4.0.0)",
    "uvicorn (>=0.34.1,<0.35.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "schemas (>=0.7.1,<0.8.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
